tests/wpt/web-platform-tests/layout-instability/sticky-descendant-move.html

<!DOCTYPE html>
<title>Layout Instability: movement of descendant of sticky positioned</title>
<link rel="help" href="https://wicg.github.io/layout-instability/" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/util.js"></script>
<div style="position: sticky; width: 400px; height: 300px; top: 0">
  <div id="child" style="position: relative; width: 300px; height: 200px; background: yellow"></div>
</div>
<script>
promise_test(async () => {
  const watcher = new ScoreWatcher;

  // Wait for the initial render to complete.
  await waitForAnimationFrames(2);

  // Modify the position of the div.
  const child = document.querySelector("#child");
  child.style.top = '100px';

  const expectedScore = computeExpectedScore(300 * (200 + 100), 100);

  // Observer fires after the frame is painted.
  assert_equals(watcher.score, 0);
  await watcher.promise;
  assert_equals(watcher.score, expectedScore);
}, 'Movement of descendant of sticky positioned.');

</script>
